// src/segment-types.ts
export interface BaseSegment {
  type: string;
  /** Indicates this segment has a definition built by the SegmentBuilder. */
  definitionKey?: string;
  negate?: boolean;
  options?: Record<string, unknown>;
}

//Pre-made segment types

/**
 * The most basic segment. Represents a segment from your data that matches a value.
 */
export interface DefaultSegment<TDefinitionKey extends string = string> extends BaseSegment {
  type: 'default';
  definitionKey?: TDefinitionKey extends readonly string[]
    ? UnionOfArrayElements<TDefinitionKey>
    : TDefinitionKey extends string
    ? TDefinitionKey
    : string;
  value: any;
}

/**
 * Represents a segment from your data that matches a  fixed time period i.e.
 * {REFERENCE_DATE} is a fixed date.
 *
 * If it were expressed in words, it would be something like:
 *
 * "All data with {FIELD} {START} {UNIT} {OPERATOR} {REFERENCE_DATE}."
 *
 * Ex: All data with "created_at" "7" "days" "before" "2021-01-01T00:00:00Z".
 */

export interface FixedTimePeriodSegment<TTimeFields extends string = string, TDefinitionKey extends string = string>
  extends BaseSegment {
  type: 'fixed_timeperiod';
  definitionKey?: TDefinitionKey extends readonly string[]
    ? UnionOfArrayElements<TDefinitionKey>
    : TDefinitionKey extends string
    ? TDefinitionKey
    : string;
  referenceDate: Date | string;
  options: {
    value: number;
    unit: string;
    field: TTimeFields;
    operator: string;
  };
}
// ...
export interface TimeRangeSegment<TDefinitionKey extends string = string> extends BaseSegment {
  type: 'time_range';
  definitionKey?: TDefinitionKey extends readonly string[]
    ? UnionOfArrayElements<TDefinitionKey>
    : TDefinitionKey extends string
    ? TDefinitionKey
    : string;
  referenceDate: Date | string;
  options: {
    start: {
      value: number;
      unit?: string;
      operatorRelativeToReference?: string;
    };
    end: {
      value: number;
      unit?: string;
      operatorRelativeToReference?: string;
    };
    field: string;
  };
}